backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pillow==10.4.0
slowapi==0.1.9

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-xdist==3.6.1
httpx==0.27.2
faker==30.8.2
requests==2.32.3
schemathesis==3.27.1

